feat: reused existing type decorator function for extensive coverage
